+++ lib/biometric_signature_method_channel.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:biometric_signature/android_config.dart';
-import 'package:biometric_signature/iOS_config.dart';
+import 'package:biometric_signature/ios_config.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
@@ -14,13 +14,15 @@ class MethodChannelBiometricSignature extends BiometricSignaturePlatform {
   final methodChannel = const MethodChannel('biometric_signature');
 
   @override
-  Future<String?> createKeys(AndroidConfig config, IOSConfig config1) async {
+  Future<String?> createKeys(
+      AndroidConfig androidConfig, IosConfig iosConfig) async {
     try {
       if (Platform.isAndroid) {
         return await methodChannel.invokeMethod<String>(
-            'createKeys', config.useStrongBox);
+            'createKeys', androidConfig.useDeviceCredentials);
       } else {
-        return await methodChannel.invokeMethod<String>('createKeys', config1.useDeviceCredentials);
+        return await methodChannel.invokeMethod<String>(
+            'createKeys', iosConfig.useDeviceCredentials);
       }
     } on PlatformException {
       rethrow;